+++ src/modules/scenarioeditor/NewScenarioView.vue
@@ -5,7 +5,7 @@
         <h1 class="text-3xl font-bold leading-tight text-gray-900">
           Create new scenario
         </h1>
-        <div class="prose-base prose mt-4">
+        <div class="prose mt-4">
           <p>
             Before you can start working on your new scenario we need some basic
             information and settings first. Remember that you can always change these
@@ -15,19 +15,44 @@
       </div>
     </header>
     <div class="mx-auto my-10 max-w-7xl sm:px-6 lg:px-8">
-      <main class="mt-6 space-y-6">
+      <form class="mt-6 space-y-6" @submit.prevent="create()">
         <FormCard
           class=""
           label="Basic scenario info"
           description="Provide a name and description for your scenario."
         >
-          <InputGroup label="Name" v-model="newScenario.name" id="name-input" />
+          <InputGroup label="Name" v-model="form.name" id="name-input" />
 
           <SimpleMarkdownInput
             label="Description"
-            v-model="newScenario.description"
+            v-model="form.description"
             description="Use markdown syntax for formatting"
           />
+        </FormCard>
+        <FormCard label="Initial ORBAT">
+          <template #description> Sides and root units.</template>
+          <div>
+            <ToggleField v-model="noInitialOrbat"
+              >Add sides and root units later</ToggleField
+            >
+          </div>
+          <template v-if="!noInitialOrbat">
+            <div
+              v-for="sideData in form.sides"
+              class="grid grid-cols-2 gap-4 rounded-md border bg-gray-50 p-4"
+            >
+              <InputGroup v-model="sideData.name" label="Side name" />
+              <SymbolCodeSelect
+                v-model="sideData.standardIdentity"
+                label="Standard identity"
+                :items="sidItems"
+              />
+              <InputGroup
+                label="Root unit name"
+                v-model="sideData.rootUnitName"
+              ></InputGroup>
+            </div>
+          </template>
         </FormCard>
         <FormCard label="Scenario start time">
           <template #description>
@@ -57,16 +82,16 @@
         </FormCard>
         <!--        <FormCard label="Order of Battle"></FormCard>-->
         <div class="flex justify-end space-x-3 px-4 sm:px-0">
-          <BaseButton primary @click="create()">Create scenario</BaseButton>
+          <BaseButton primary type="submit">Create scenario</BaseButton>
           <BaseButton @click="cancel()">Cancel</BaseButton>
         </div>
-      </main>
+      </form>
     </div>
   </div>
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
+import { reactive, ref } from "vue";
 import FormCard from "../../components/FormCard.vue";
 import InputGroup from "../../components/InputGroup.vue";
 import SimpleMarkdownInput from "../../components/SimpleMarkdownInput.vue";
@@ -78,6 +103,16 @@ import { useRouter } from "vue-router";
 import { SCENARIO_ROUTE } from "@/router/names";
 import { useScenario } from "@/scenariostore";
 import { createEmptyScenario } from "@/scenariostore/io";
+import ToggleField from "@/components/ToggleField.vue";
+import SymbolCodeSelect from "@/components/SymbolCodeSelect.vue";
+import { sidItems } from "@/symbology/helpers";
+import { ScenarioInfo, SideData } from "@/types/scenarioModels";
+import { SID } from "@/symbology/values";
+import { nanoid } from "@/utils";
+import { Sidc } from "@/symbology/sidc";
+
+const router = useRouter();
+const { scenario } = useScenario();
 
 const standardSettings = [
   {
@@ -92,26 +127,61 @@ const standardSettings = [
   },
 ];
 
+interface InitialSideData extends SideData {
+  rootUnitName?: string;
+  rootUnitSidc?: string;
+}
+
+interface NewScenarioForm extends ScenarioInfo {
+  sides: InitialSideData[];
+}
+
+const noInitialOrbat = ref(false);
+
 const newScenario = ref(createEmptyScenario());
-
-const { scenario } = useScenario();
-
-const router = useRouter();
-
 const timeZone = ref(newScenario.value.timeZone || "UTC");
-
 const { year, month, day, hour, minute, resDateTime } = useYMDElements({
   timestamp: newScenario.value.startTime!,
   isLocal: true,
   timeZone,
 });
 
+const form = reactive<NewScenarioForm>({
+  name: "New scenario",
+  description: "Scenario description",
+  sides: [
+    { name: "Side 1", standardIdentity: SID.Friend, rootUnitName: "HQ" },
+    { name: "Side 2", standardIdentity: SID.Hostile, rootUnitName: "HQ" },
+  ],
+});
+
 function create() {
   const startTime = resDateTime.value.valueOf();
   newScenario.value.startTime = startTime;
+  newScenario.value.name = form.name;
+  newScenario.value.description = form.description;
   scenario.value.io.loadFromObject(newScenario.value);
   scenario.value.time.setCurrentTime(startTime);
-
+  const { state } = scenario.value.store;
+  const { unitActions } = scenario.value;
+  if (!noInitialOrbat.value) {
+    form.sides.forEach((sideData) => {
+      const sideId = unitActions.addSide(sideData, { markAsNew: false });
+      const parentId = state.getSideById(sideId).groups[0];
+      const sidc = new Sidc("10031000000000000000");
+      sidc.standardIdentity = sideData.standardIdentity;
+      unitActions.addUnit(
+        {
+          id: nanoid(),
+          name: sideData.rootUnitName ?? "test",
+          sidc: sidc.toString(),
+          subUnits: [],
+          _pid: "nn",
+        },
+        parentId
+      );
+    });
+  }
   router.push({ name: SCENARIO_ROUTE });
 }